rss2适配频道

+++ rss2/del_dy.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 
 from nonebot import on_command, CommandSession
-from nonebot.permission import *
+from .permission import admin_permission
 
 from .RSS import my_trigger as tr
 from .RSS import rss_class
@@ -11,32 +11,47 @@ from .RSS import rss_class
 FILE_PATH = str(str(Path.cwd()) + os.sep + "data" + os.sep)
 
 
-@on_command("deldy", aliases=("drop", "删除订阅"), permission=GROUP_ADMIN | SUPERUSER)
+@on_command(
+    "deldy", aliases=("drop", "删除订阅"), permission=admin_permission, only_to_me=False
+)
 async def deldy(session: CommandSession):
     rss_name = session.get("deldy", prompt="输入要删除的订阅名")
     group_id = session.ctx.get("group_id")
+    guild_channel_id = session.ctx.get("guild_id")
+    if guild_channel_id:
+        group_id = None
+        guild_channel_id = guild_channel_id + "@" + session.ctx.get("channel_id")
 
     rss = rss_class.Rss()
-    if rss.find_name(name=rss_name):
-        rss = rss.find_name(name=rss_name)
-    else:
-        await session.send("❌ 删除失败！不存在该订阅！")
-        return
+    rss = rss.find_name(name=rss_name)
 
-    if group_id:
-        if rss.delete_group(group=group_id):
-            if not rss.group_id and not rss.user_id:
-                rss.delete_rss()
-                await tr.delete_job(rss)
-            else:
-                await tr.add_job(rss)
-            await session.send(f"👏 当前群组取消订阅 {rss.name} 成功！")
-        else:
-            await session.send(f"❌ 当前群组没有订阅： {rss.name} ！")
+    if rss is None:
+        await session.finish("❌ 删除失败！不存在该订阅！")
     else:
-        rss.delete_rss()
-        await tr.delete_job(rss)
-        await session.send(f"👏 订阅 {rss.name} 删除成功！")
+        if group_id:
+            if rss.delete_group(group=str(group_id)):
+                if not rss.group_id and not rss.user_id and not rss.guild_channel_id:
+                    rss.delete_rss()
+                    await tr.delete_job(rss)
+                else:
+                    await tr.add_job(rss)
+                await session.finish(f"👏 当前群组取消订阅 {rss.name} 成功！")
+            else:
+                await session.finish(f"❌ 当前群组没有订阅： {rss.name} ！")
+        elif guild_channel_id:
+            if rss.delete_guild_channel(guild_channel=guild_channel_id):
+                if not rss.group_id and not rss.user_id and not rss.guild_channel_id:
+                    rss.delete_rss()
+                    await tr.delete_job(rss)
+                else:
+                    await tr.add_job(rss)
+                await session.finish(f"👏 当前子频道取消订阅 {rss.name} 成功！")
+            else:
+                await session.finish(f"❌ 当前子频道没有订阅： {rss.name} ！")
+        else:
+            rss.delete_rss()
+            await tr.delete_job(rss)
+            await session.finish(f"👏 订阅 {rss.name} 删除成功！")
 
 
 @deldy.args_parser
